+++ HPXMLtoOpenStudio/measure.rb
@@ -233,6 +233,7 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
 
       # Output
       season_day_nums = add_unmet_hours_output(model, hpxml_osm_map, hpxml)
+      add_unmet_loads_output(model, hpxml_osm_map)
       loads_data = add_total_loads_output(model, hpxml_osm_map)
       if args[:add_component_loads]
         add_component_loads_output(model, hpxml_osm_map, loads_data, season_day_nums)
@@ -1796,10 +1797,10 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     end
 
     # Hot water fixtures and appliances
-    HotWaterAndAppliances.apply(model, runner, @hpxml_header, @hpxml_bldg, weather, spaces, hot_water_distribution,
-                                solar_thermal_system, @eri_version, @schedules_file, plantloop_map,
-                                @hpxml_header.unavailable_periods, @hpxml_bldg.building_construction.number_of_units,
-                                @apply_ashrae140_assumptions)
+    showers_peak_flows = HotWaterAndAppliances.apply(model, runner, @hpxml_header, @hpxml_bldg, weather, spaces, hot_water_distribution,
+                                                     solar_thermal_system, @eri_version, @schedules_file, plantloop_map,
+                                                     @hpxml_header.unavailable_periods, @hpxml_bldg.building_construction.number_of_units,
+                                                     @apply_ashrae140_assumptions)
 
     if (not solar_thermal_system.nil?) && (not solar_thermal_system.collector_area.nil?) # Detailed solar water heater
       loc_space, loc_schedule = get_space_or_schedule_from_location(solar_thermal_system.water_heating_system.location, model, spaces)
@@ -1808,6 +1809,9 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
 
     # Add combi-system EMS program with water use equipment information
     Waterheater.apply_combi_system_EMS(model, @hpxml_bldg.water_heating_systems, plantloop_map)
+
+    # Add unmet wh loads calculation
+    Waterheater.unmet_wh_loads_program(model, @hpxml_bldg.water_heating_systems, plantloop_map, showers_peak_flows)
   end
 
   # TODO
@@ -2587,6 +2591,50 @@ class HPXMLtoOpenStudio < OpenStudio::Measure::ModelMeasure
     program_calling_manager.addProgram(program)
 
     return season_day_nums
+  end
+
+  # TODO
+  #
+  # @param model [OpenStudio::Model::Model] OpenStudio Model object
+  # @param hpxml_osm_map [Hash] Map of HPXML::Building objects => OpenStudio Model objects for each dwelling unit
+  # @return [void]
+  def add_unmet_loads_output(model, hpxml_osm_map)
+    return if hpxml_osm_map.select { |hpxml_bldg, _unit_model| !hpxml_bldg.water_heating_systems.empty? }.empty?
+
+    # Retrieve objects
+    shower_e_vars = []
+    shower_sag_time_vars = []
+    unit_multipliers = []
+    hpxml_osm_map.each do |hpxml_bldg, unit_model|
+      shower_e_vars << unit_model.getEnergyManagementSystemGlobalVariables.find { |v| v.additionalProperties.getFeatureAsString('ObjectType').to_s == Constants.ObjectNameUnmetLoadsShowerE }
+      shower_sag_time_vars << unit_model.getEnergyManagementSystemGlobalVariables.find { |v| v.additionalProperties.getFeatureAsString('ObjectType').to_s == Constants.ObjectNameUnmetLoadsShowerSagTime }
+      unit_multipliers << hpxml_bldg.building_construction.number_of_units
+    end
+
+    # EMS program
+    total_shower_e = 'total_shower_e'
+    total_shower_sag_time = 'total_shower_sag_time'
+    unit_shower_sag_time = 'unit_shower_sag_time'
+    program = OpenStudio::Model::EnergyManagementSystemProgram.new(model)
+    program.setName('unmet loads program')
+    program.additionalProperties.setFeature('ObjectType', Constants.ObjectNameUnmetLoadsProgram)
+    program.addLine("Set #{total_shower_e} = 0")
+    program.addLine("Set #{total_shower_sag_time} = 0")
+    shower_e_vars.each_with_index do |shower_e_var, i|
+      program.addLine("Set total_shower_e = total_shower_e + (#{shower_e_var.name} * #{unit_multipliers[i]})")
+    end
+    shower_sag_time_vars.each_with_index do |shower_sag_time_var, _i|
+      program.addLine("Set #{unit_shower_sag_time} = #{shower_sag_time_var.name}")
+      program.addLine("  If #{unit_shower_sag_time} > #{total_shower_sag_time}") # Use max hourly value across all units
+      program.addLine("    Set #{total_shower_sag_time} = #{unit_shower_sag_time}")
+      program.addLine('  EndIf')
+    end
+
+    # EMS calling manager
+    manager = OpenStudio::Model::EnergyManagementSystemProgramCallingManager.new(model)
+    manager.setName("#{program.name} calling manager")
+    manager.setCallingPoint('EndOfZoneTimestepAfterZoneReporting')
+    manager.addProgram(program)
   end
 
   # TODO