+++ Changelog.md
@@ -18,6 +18,7 @@ __New Features__
   - **Breaking change**: Replaced `slab_perimeter_depth` argument with `slab_perimeter_insulation_depth`.
 - **Breaking change**: Disaggregates "Walls" into "Above Grade Walls" and "Below Grade Walls" in results_design_load_details.csv output file.
 - Adds a warning if the sum of supply/return duct leakage to outside values is very high.
+- Adds metrics for unmet hot water loads.
 
 __Bugfixes__
 - Prevents possible error when using multiple `Attic`/`Foundation` elements for the same attic/foundation type.